# src/main/openapi/bus-ticket-booking-app.yml
openapi: 3.1.0
info:
  title: Bus Ticket Booking API
  description: API for managing bus ticket booking application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/routes:
    get:
      tags:
        - Route Management
      summary: Get all routes
      description: Retrieve a list of all routes available in the system.
      operationId: GetAllRoutes
      responses:
        "200":
          description: Successfully retrieved all routes
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/RouteResponse"
        "403":
          description: Forbidden - User does not have ADMIN role
          content:
            application/json:
              schema:
                type: string
                example: Access denied
        "404":
          description: No routes found
          content:
            application/json:
              schema:
                type: string
                example: No routes available
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
    post:
      tags:
        - Route Management
      summary: Create a new route
      description: Allows users with ADMIN role to create a new route.
      operationId: CreateRoute
      requestBody:
        description: Route details to be created
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
        required: true
      responses:
        "201":
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          description: Invalid route request data
          content:
            application/json:
              schema:
                type: string
                example: Invalid route request data
        "403":
          description: Forbidden - User does not have ADMIN role
          content:
            application/json:
              schema:
                type: string
                example: Access denied
        "409":
          description: Conflict - Route already exists
          content:
            application/json:
              schema:
                type: string
                example: Route already exists
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
...
  /api/v1/routes/{id}/base-price:
    patch:
      tags:
        - Route Management
      summary: Update base price of a route
      description: Allows users with ADMIN role to update the base price of a specific
        route.
      operationId: UpdateBasePrice
      parameters:
        - name: id
          in: path
          description: ID of the route to update
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        description: New base price for the route
        content:
          application/json:
            schema:
              type: number
              format: decimal
              example: 100.0
        required: true
      responses:
        "200":
          description: Base price updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          description: Invalid base price value
          content:
            application/json:
              schema:
                type: string
                example: Base price must be greater than 0
        "403":
          description: Forbidden - User does not have ADMIN role
          content:
            application/json:
              schema:
                type: string
                example: Access denied
        "404":
          description: Route not found
          content:
            application/json:
              schema:
                type: string
                example: Route with the specified ID does not exist
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
  /api/v1/seats/{scheduled-trip-id}:
    get:
      tags:
        - Seat Management
      summary: Get seats for a scheduled trip
      description: Retrieve a list of seats available for a specific scheduled trip.
      operationId: GetSeatsForScheduledTrip
      parameters:
        - name: scheduled-trip-id
          in: path
          description: ID of the scheduled trip for which to retrieve seats
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Successfully retrieved seats for the scheduled trip
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/SeatResponse"
        "404":
          description: Scheduled trip not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeatResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
...
components:
  schemas:
...
    RouteRequest:
      type: object
      description: Request object for creating or updating a route
      properties:
        origin:
          type: string
          description: Origin city of the route
          example: Prague
          minLength: 1
        destination:
          type: string
          description: Destination city of the route
          example: Vienna
          minLength: 1
        distance:
          type: number
          format: double
          description: Distance of the route in kilometers
          example: 250.0
          minimum: 0
        duration:
          type: string
          description: Duration in hours and minutes
          example: 02:30
          pattern: "^(\\d{2}:(\\d{2})$"
        basePriceCzk:
          type: number
          description: Base price of the route in CZK
          example: 300.0
          minimum: 0
      required:
        - basePriceCzk
        - destination
        - distance
        - duration
        - origin
    RouteResponse:
      type: object
      description: Response object for route details
      properties:
        origin:
          type: string
          description: Origin of the route
          example: Prague
        destination:
          type: string
          description: Destination of the route
          example: Vienna
        distance:
          type: number
          format: double
          description: Distance of the route in kilometers
          example: 250.0
        duration:
          type: string
          description: Duration in hours and minutes
          example: PT2H30M
          pattern: "^(PT)(\\d{1,2})H(\\d{1,2})(M)"
        basePriceCzk:
          type: number
          description: Base price of the route in CZK
          example: 300.0
      required:
        - basePriceCzk
        - destination
        - distance
        - duration
        - origin
...
    SeatResponse:
      type: object
      description: Response object for seat details
      properties:
        seatNumber:
          type: integer
          format: int32
          description: Unique identifier for the seat
          example: 12
        status:
          type: string
          description: Current status of the seat
          enum:
            - FREE
            - BLOCKED
            - RESERVED
      required:
        - seatNumber
        - status
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT